Refactor Integration Tests to use plugins block
By using the plugins block we can test the actual usage of most users
(since the buildscript approach is unfavored and provides less
features). However, since our plugin is not really published (or at
least the versions we will be testing aren't published once it actually
gets published for the first time) we need to bypass the plugin
resolution to redirect to the shadow jar that was built.

Using the plugins block also means that we can't use a SNAPSHOT version,
so the version was changed to 0.1.0 in order to be accepted by Gradle on
the Integration Tests.

Finally, since our plugin is loaded through the plugins block then the
accessor for the Kotlin DSL version is automatically generated, so we no
longer require the publicly available extension functions on Project.

Lastly, the integration tests were bundled into a single task called
'integrationTests' so that it can be easily ignored when developing
locally.